+++ CHANGES.md
@@ -23,6 +23,7 @@ To be released.
     `DetectLanguagePathnameOptions` interface, and
     `DetectLanguageSearchParamsOptions` interface to _path.ts_ module.
  -  Renamed `Resource#hasRepresentation()` method to `has()`.
+ -  Renamed `Resource#getRepresentation()` method to `get()`.
 
 
 Version 0.1.1